suicidolino

+++ test/main.c
@@ -125,7 +125,7 @@ int		is_in_line(int x, int y, t_coord A, t_coord B)
 		return ((x == B.X) && (y == B.Y));
 	else if (!(B.X - A.X) || !(B.Y - A.Y))
 		return (0);
-	return (((B.X - A.X) * (y - A.Y)) == ((x - A.X) * (B.Y - A.Y)));
+	return ((int)((B.X - A.X) * (y - A.Y)) == (int)((x - A.X) * (B.Y - A.Y)));
 }
 
 int		is_in_lines(int x, int y, t_cupos cube)
